requirements.txt
astroid==2.6.6
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cssselect2==0.4.1
defusedxml==0.7.1
idna==3.2
iqoptionapi @ git+git://github.com/iqoptionapi/iqoptionapi.git@02dd4790266b30efff9f3d49bfc4b9815dd9858f
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.3.1
pycparser==2.20
pylint==2.9.6
PySide2==5.15.2
PySimpleGUI==4.45.0
PySimpleGUIQt==0.35.0
requests==2.26.0
shiboken2==5.15.2
six==1.16.0
tinycss2==1.1.0
toml==0.10.2
urllib3==1.26.6
webencodings==0.5.1
websocket-client==0.56.0
wrapt==1.12.1
